+++ app/Http/Controllers/CodeController.php
@@ -486,6 +486,8 @@ class CodeController extends Controller
             ->timeout(10)
             ->command($compilerCommand)->run();
         
+        $log->info("compiler exited with code: " . $compilerProcessResult->exitCode());
+
         if($compilerProcessResult->exitCode() !== 0)
         {
             $response = [
@@ -554,14 +556,22 @@ class CodeController extends Controller
     
     function filterOutput($text)
     {
-        $text = array_filter(explode("\n", $text), function($value)
-        {
-            return (strpos($value, "undefined symbol") !== false) ||
-                (strpos($value, "duplicate symbol") !== false) ||
-                (strpos($value, "pgetinker.cpp") === 0);
-        });
+        
+        $text = str_replace("/opt/emsdk/upstream/emscripten/cache/sysroot", "/***", $text);
 
-        return implode("\n", $text);
+        return $text;
+        // $text = explode("\n", $text);
+
+        // for($i = 0 $i < count($text))
+
+        // // $text = array_filter(explode("\n", $text), function($value)
+        // // {
+        // //     return (strpos($value, "undefined symbol") !== false) ||
+        // //         (strpos($value, "duplicate symbol") !== false) ||
+        // //         (strpos($value, "pgetinker.cpp") === 0);
+        // // });
+        
+        // return implode("\n", $text);
     }
 
 }